Move from submit page to a submit modal
