+++ package.json
@@ -21,12 +21,14 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-helmet": "^6.1.0"
+    "react-helmet": "^6.1.0",
+    "react-modal": "^3.14.2"
   },
   "devDependencies": {
     "@babel/core": "^7.14.0",
     "@babel/preset-react": "^7.13.13",
     "@babel/preset-typescript": "^7.13.0",
+    "@types/react-modal": "^3.12.0",
     "autoprefixer": "^10.2.5",
     "babel-plugin-module-resolver": "^4.1.0",
     "babel-preset-gatsby": "^1.4.0",
