+++ src/components/layout/Layout.tsx
@@ -1,6 +1,8 @@
-import React, { ReactNode } from "react"
+import React, { ReactNode, useState } from "react"
 import Header from "./Header"
 import Footer from "./Footer"
+import Modal from "react-modal"
+import SubmitModal from "./SubmitModal"
 
 interface Props {
   children: ReactNode
@@ -17,19 +19,24 @@ export const Layout: React.FC<Props> = ({
   hideSearchBar,
   headerText,
   subheaderText,
-}) => (
-  <div>
-    <header>
-      <Header
-        hideSearchBar={hideSearchBar}
-        hideSubmitButton={hideSubmitButton}
-        headerText={headerText}
-        subheaderText={subheaderText}
-      />
-    </header>
-    <main className="px-12 py-6">{children}</main>
-    <footer>
-      <Footer />
-    </footer>
-  </div>
-)
+}) => {
+  const [modalOpen, setModalOpen] = useState<boolean>(false)
+  return (
+    <div>
+      <header>
+        <Header
+          onSubmitPress={() => setModalOpen(true)}
+          hideSearchBar={hideSearchBar}
+          hideSubmitButton={hideSubmitButton}
+          headerText={headerText}
+          subheaderText={subheaderText}
+        />
+      </header>
+      <main className="px-12 py-6">{children}</main>
+      <footer>
+        <Footer />
+      </footer>
+      <SubmitModal open={modalOpen} onClose={() => setModalOpen(false)} />
+    </div>
+  )
+}
